src/parsing/pdf.py
"""src/parsing/pdf.py
###############################################################################
PDF text-extraction adapter
###############################################################################
The helper in this module converts a *binary* PDF payload obtained via
`starlette.datastructures.UploadFile` into a UTF-8 ``str`` suitable for the
classification pipeline.

Design-notes
============
• **Non-blocking** – `pdfminer.six` is CPU-bound and synchronous.  We therefore
  off-load processing to a background thread with :pyfunc:`asyncio.to_thread`.
• **Single responsibility** – exactly one public coroutine ≤ 40 lines of code.
• **No catch-all** – let `pdfminer` propagate domain-specific exceptions (e.g.
  `PDFSyntaxError`) so up-stack handlers can deal with them.  We *only* add a
  debug log entry on success for observability.
• **Memory considerations** – the PDF is already ≤ 10 MB (validated earlier).
  Reading it fully into memory is acceptable and cheaper than multiple I/O
  passes that `pdfminer` would otherwise perform on a temporary file.
"""

from __future__ import annotations

# stdlib
import asyncio
from io import BytesIO
from typing import Final

# third-party
import structlog
from pdfminer.high_level import extract_text  # type: ignore[import-not-found]
from starlette.datastructures import UploadFile

__all__: list[str] = ["extract_text_from_pdf"]

logger = structlog.get_logger(__name__)

# pdfminer default codec is latin-1; we always re-encode to UTF-8 for internal use
_PDFMINER_CODEC: Final[str] = "utf-8"


async def extract_text_from_pdf(file: UploadFile) -> str:
    """
    Extract text content from a PDF file using pdfminer.six.

    Args:
        file: The uploaded PDF file

    Returns:
        Extracted text content as a string
    """
    # Read the file content
    await file.seek(0)
    content = await file.read()

    # Define worker function for async execution
    def _worker(pdf_content: bytes) -> str:
        # Create BytesIO object for pdfminer to read from
        pdf_buffer = BytesIO(pdf_content)

        try:
            # Extract text using pdfminer
            extracted_text = extract_text(pdf_buffer)
            return extracted_text or ""
        except Exception:
            # Handle PDF parsing errors silently
            return ""

    # Run CPU-bound extraction in threadpool
    return await asyncio.to_thread(_worker, content)
