requirements.txt
# -----------------------------------------------------------------------------
# Runtime ─ Web & Core
# -----------------------------------------------------------------------------
fastapi
flask
uvicorn[standard]

# -----------------------------------------------------------------------------
# Config & Validation
# -----------------------------------------------------------------------------
python-dotenv
pydantic
pydantic-settings
python-multipart

# -----------------------------------------------------------------------------
# Parsing & ML
# -----------------------------------------------------------------------------
pdfminer.six
docx2txt
pandas
scikit-learn
pytesseract
Pillow

# -----------------------------------------------------------------------------
# Logging & Observability
# -----------------------------------------------------------------------------
structlog
python-json-logger
prometheus-fastapi-instrumentator
PyYAML

# -----------------------------------------------------------------------------
# Development / Tooling
# -----------------------------------------------------------------------------
black
ruff
mypy
pre-commit
isort
requests

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest
pytest-asyncio
pytest-cov
pytest-mock
faker
coverage
httpx
